Issue #951: extract tau-access crate
