+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "crates/tau-session",
   "crates/tau-skills",
   "crates/tau-extensions",
+  "crates/tau-access",
   "crates/tau-orchestrator",
   "crates/tau-runtime",
   "crates/tau-tui",
