+++ crates/tau-access/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod approvals;
+pub mod pairing;
+pub mod rbac;
+pub mod trust_roots;
+
+pub use approvals::*;
+pub use pairing::*;
+pub use rbac::*;
+pub use trust_roots::*;
